pyproject.toml
[project]
name = "speech-speaker-recognition"
version = "0.1.0"
description = "Speech recognition and speaker diarization pipeline using Whisper and pyannote"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    # PyTorch still needed for pyannote.audio
    "torch>=2.0.0",
    "torchaudio>=2.0.0",

    # Faster Whisper implementation (2-4x faster, same accuracy)
    "faster-whisper>=1.0.0",
    "pyannote.audio>=3.1.0",
    "pyannote.core>=5.0.0",
    
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "tqdm>=4.65.0",
    
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "pydub>=0.25.0",
    
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "matplotlib>=3.7.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["speech_pipeline"]

[project.scripts]
speech-pipeline = "speech_pipeline.cli:main"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true